+++ plugin-dev/Source/SentryEditor/Public/SentryEditorModule.h
@@ -1,0 +1,14 @@
+﻿// Copyright (c) 2022 Sentry. All Rights Reserved.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "Modules/ModuleInterface.h"
+
+class FSentryEditorModule : public IModuleInterface
+{
+public:
+	/** IModuleInterface implementation */
+	virtual void StartupModule() override;
+	virtual void ShutdownModule() override;
+};